extract kyma related test util funcs to separate file
