+++ controllers/manifest_controller/manifest_controller_test.go
@@ -1,7 +1,6 @@
 package manifest_controller_test
 
 import (
-	"errors"
 	"os"
 	"path/filepath"
 	"strings"
@@ -14,8 +13,6 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-var ErrManifestStateMisMatch = errors.New("ManifestState mismatch")
 
 var _ = Describe(
 	"Rendering manifest install layer", func() {
