+++ pkg/testutils/utils.go
@@ -53,40 +53,6 @@ var (
 	ErrDeletionTimestampFound = errors.New("deletion timestamp not nil")
 )
 
-func NewTestKyma(name string) *v1beta2.Kyma {
-	return newKCPKymaWithNamespace(name, v1.NamespaceDefault, v1beta2.DefaultChannel, v1beta2.SyncStrategyLocalClient)
-}
-
-func NewKymaForE2E(name, namespace, channel string) *v1beta2.Kyma {
-	kyma := newKCPKymaWithNamespace(name, namespace, channel, v1beta2.SyncStrategyLocalSecret)
-	kyma.Labels[v1beta2.SyncLabel] = v1beta2.EnableLabelValue
-	return kyma
-}
-
-func newKCPKymaWithNamespace(name, namespace, channel, syncStrategy string) *v1beta2.Kyma {
-	return &v1beta2.Kyma{
-		TypeMeta: v1.TypeMeta{
-			APIVersion: v1beta2.GroupVersion.String(),
-			Kind:       string(v1beta2.KymaKind),
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name:      fmt.Sprintf("%s-%s", name, RandomName()),
-			Namespace: namespace,
-			Annotations: map[string]string{
-				watcher.DomainAnnotation:       "example.domain.com",
-				v1beta2.SyncStrategyAnnotation: syncStrategy,
-			},
-			Labels: map[string]string{
-				v1beta2.InstanceIDLabel: "test-instance",
-			},
-		},
-		Spec: v1beta2.KymaSpec{
-			Modules: []v1beta2.Module{},
-			Channel: channel,
-		},
-	}
-}
-
 func NewTestManifest(prefix string) *v1beta2.Manifest {
 	return &v1beta2.Manifest{
 		ObjectMeta: v1.ObjectMeta{
@@ -220,50 +186,6 @@ func CreateCR(ctx context.Context, clnt client.Client, obj client.Object) error 
 		return err
 	}
 	return nil
-}
-
-func SyncKyma(ctx context.Context, clnt client.Client, kyma *v1beta2.Kyma) error {
-	err := clnt.Get(ctx, client.ObjectKey{
-		Name:      kyma.Name,
-		Namespace: kyma.Namespace,
-	}, kyma)
-	// It might happen in some test case, kyma get deleted, if you need to make sure Kyma should exist,
-	// write expected condition to check it specifically.
-	return client.IgnoreNotFound(err)
-}
-
-func KymaExists(ctx context.Context, clnt client.Client, name, namespace string) error {
-	kyma, err := GetKyma(ctx, clnt, name, namespace)
-	if util.IsNotFound(err) {
-		return ErrNotFound
-	}
-	if kyma != nil && kyma.DeletionTimestamp != nil {
-		return ErrDeletionTimestampFound
-	}
-	return nil
-}
-
-func GetKyma(ctx context.Context, testClient client.Client, name, namespace string) (*v1beta2.Kyma, error) {
-	kymaInCluster := &v1beta2.Kyma{}
-	if namespace == "" {
-		namespace = v1.NamespaceDefault
-	}
-	err := testClient.Get(ctx, client.ObjectKey{
-		Namespace: namespace,
-		Name:      name,
-	}, kymaInCluster)
-	if err != nil {
-		return nil, err
-	}
-	return kymaInCluster, nil
-}
-
-func IsKymaInState(ctx context.Context, kcpClient client.Client, kymaName string, state v1beta2.State) bool {
-	kymaFromCluster, err := GetKyma(ctx, kcpClient, kymaName, "")
-	if err != nil || kymaFromCluster.Status.State != state {
-		return false
-	}
-	return true
 }
 
 func GetManifestSpecRemote(
@@ -466,26 +388,6 @@ func AppendExternalCRDs(path string, files ...string) []*apiExtensionsv1.CustomR
 		}
 	}
 	return crds
-}
-
-func ExpectKymaManagerField(
-	ctx context.Context, controlPlaneClient client.Client, kymaName string, managerName string,
-) (bool, error) {
-	createdKyma, err := GetKyma(ctx, controlPlaneClient, kymaName, "")
-	if err != nil {
-		return false, err
-	}
-	if createdKyma.ManagedFields == nil {
-		return false, nil
-	}
-
-	for _, v := range createdKyma.ManagedFields {
-		if v.Subresource == "status" && v.Manager == managerName {
-			return true, nil
-		}
-	}
-
-	return false, nil
 }
 
 func GetModuleTemplate(ctx context.Context,
