+++ tests/e2e_test/kyma_metrics_test.go
@@ -2,14 +2,14 @@ package e2e_test
 
 import (
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
-	"github.com/kyma-project/lifecycle-manager/pkg/testutils"
+	. "github.com/kyma-project/lifecycle-manager/pkg/testutils"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"k8s.io/apimachinery/pkg/api/meta"
 )
 
 var _ = Describe("Kyma Metrics", Ordered, func() {
-	kyma := testutils.NewKymaForE2E("kyma-sample", "kcp-system", "regular")
+	kyma := NewKymaForE2E("kyma-sample", "kcp-system", "regular")
 	GinkgoWriter.Printf("kyma before create %v\n", kyma)
 
 	BeforeAll(func() {
@@ -53,7 +53,7 @@ var _ = Describe("Kyma Metrics", Ordered, func() {
 			Should(Succeed())
 
 		By("waiting for Kyma CR to be removed")
-		Eventually(CheckKCPKymaCRDeleted).
+		Eventually(KymaDeleted).
 			WithContext(ctx).
 			WithArguments(kyma.GetName(), kyma.GetNamespace(), controlPlaneClient).
 			Should(Succeed())
